feat(search): Add search page
When adding search page, some hooks(useDebounce) and actions were needed.
Therefore, some documents supporting the search page were made including the above.

P.S. As the concept was changed from the game-lesson platform into the movie review,
some parts in the menu of Navbar, hooks and components' contents were changed accordingly.
The Logo was changed as well.
In addition, the year on the footer was changed from 2021 into 2022.